renamed the experiment to gcp-vm-instance-stop
