+++ experiments/gcp/gcp-vm-instance-stop/rbac.yaml
@@ -2,18 +2,18 @@
 apiVersion: v1
 kind: ServiceAccount
 metadata:
-  name: vm-instance-stop-sa
+  name: gcp-vm-instance-stop-sa
   namespace: default
   labels:
-    name: vm-instance-stop-sa
+    name: gcp-vm-instance-stop-sa
     app.kubernetes.io/part-of: litmus
 ---
 apiVersion: rbac.authorization.k8s.io/v1
 kind: ClusterRole
 metadata:
-  name: vm-instance-stop-sa
+  name: gcp-vm-instance-stop-sa
   labels:
-    name: vm-instance-stop-sa
+    name: gcp-vm-instance-stop-sa
     app.kubernetes.io/part-of: litmus
 rules:
 - apiGroups: [""]
@@ -35,15 +35,15 @@ rules:
 apiVersion: rbac.authorization.k8s.io/v1
 kind: ClusterRoleBinding
 metadata:
-  name: vm-instance-stop-sa
+  name: gcp-vm-instance-stop-sa
   labels:
-    name: vm-instance-stop-sa
+    name: gcp-vm-instance-stop-sa
     app.kubernetes.io/part-of: litmus
 roleRef:
   apiGroup: rbac.authorization.k8s.io
   kind: ClusterRole
-  name: vm-instance-stop-sa
+  name: gcp-vm-instance-stop-sa
 subjects:
 - kind: ServiceAccount
-  name: vm-instance-stop-sa
+  name: gcp-vm-instance-stop-sa
   namespace: default
