+++ experiments/gcp/gcp-vm-instance-stop/test/test.yml
@@ -13,7 +13,7 @@ spec:
       labels: 
         app: litmus-experiment
     spec:
-      serviceAccountName: vm-instance-stop-sa
+      serviceAccountName: gcp-vm-instance-stop-sa
       containers:
       - name: gotest
         image: busybox
