+++ pkg/gcp/gcp-vm-instance-stop/types/types.go
